Add files via upload
In this update the store looks the same. But once you enter the Items shop the items have changed to include the price, and a quantity selector. once something is added to the cart, and the user loads the cart page the change are the total price is displayed and the total number of items and each item is also displayed under in this format( Sulfuric Acid 1 Grams ( $21.99 Per Gram) Total = 21.99).